pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "python-dotenv>=1.0.1",
    "openai-agents>=0.0.15",
    "sendgrid>=6.11.0",
    "setuptools"
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
